+++ variants/cardputer_adv/companion_radio/ui-cardputer/screen/SettingsScreen.h
@@ -20,6 +20,7 @@ class SettingsScreen : public CardputerScreen {
     DeviceBatteryCorrection,
     HdrMesh,
     MeshPathSize,
+    MeshDefaultScope,
     Count
   };